Restore the map-loading status return type and native loop
IsLoadMapCompleted returns a scalar status in r3; the retail script caller
reads that register. Replace the Ghidra-derived 64-bit signature and fabricated
high/low return words with int, and check the 16 async handles using an ordinary
array loop instead of manually unrolled CMapMng pointer casts.

In the script caller, retain the map status before checking the part status,
matching the retail evaluation order and removing the obsolete return cast.
The six instructions consuming the map result and calling the part check now
match retail exactly.

The native loop compiles to the same 65 instructions as retail, with only 19
consistent r4/r5 operand swaps. Accept the score change from 100% to 97.84615%
(and 260 fewer exact bytes) to recover the correct ABI and plausible source.
The script dispatcher improves from 97.460785% to 97.46247%; no other function
scores change. Fill the EN map annotation for the status method.

Validation: full source build, whole report comparison, affected-unit objdiff,
consistent-register-renaming check, exact caller sequence comparison,
git diff --check, and PAL SHA-1 check.

+++ include/ffcc/p_map.h
@@ -21,7 +21,7 @@ public:
     void create();
     void createViewer();
     void LoadMap(int, int, void*, unsigned long, unsigned char);
-    unsigned long long IsLoadMapCompleted();
+    int IsLoadMapCompleted();
     void destroy();
     void calcInit();
     void calc();
